+++ src/main/java/net/brnbrd/delightful/common/item/food/Nutrition.java
@@ -69,11 +69,11 @@ public class Nutrition {
     public static final FoodProperties PRICKLY_PEAR_JUICE = (new FoodProperties.Builder())
         .effect(() -> new MobEffectInstance(MobEffects.SATURATION, 400, 0), 1.0F).build();
     public static final FoodProperties LAVENDER_TEA = (new FoodProperties.Builder()).alwaysEat()
-        .effect(() -> new MobEffectInstance(ModEffects.COMFORT.get(), 1200, 0), 1.0F)
-        .effect(() -> new MobEffectInstance(Mods.getCaffeinated().get(), 200, 0), 1.0F).build();
+        .effect(() -> new MobEffectInstance(ModEffects.COMFORT.get(), 600, 0), 1.0F)
+        .effect(() -> new MobEffectInstance(Mods.getCaffeinated().get(), 300, 0), 1.0F).build();
     public static final FoodProperties AZALEA_TEA = (new FoodProperties.Builder()).alwaysEat()
-        .effect(() -> new MobEffectInstance(ModEffects.NOURISHMENT.get(), 1200, 0), 1.0F)
-        .effect(() -> new MobEffectInstance(Mods.getCaffeinated().get(), 200, 0), 1.0F).build();
+        .effect(() -> new MobEffectInstance(ModEffects.NOURISHMENT.get(), 900, 0), 1.0F)
+        .effect(() -> new MobEffectInstance(Mods.getCaffeinated().get(), 300, 0), 1.0F).build();
     public static final FoodProperties MATCHA_LATTE = (new FoodProperties.Builder()).alwaysEat().nutrition(6).saturationMod(0.5F)
         .effect(() -> new MobEffectInstance(Mods.getGreenTeaEffect().get(), 1200, 1), 1.0F)
         .effect(() -> new MobEffectInstance(Mods.getCaffeinated().get(), 1200, 0), 1.0F).build();