+++ src/main/java/net/brnbrd/delightful/data/gen/DelightfulEntityTagProvider.java
@@ -9,6 +9,8 @@ import net.minecraft.data.PackOutput;
 import net.minecraft.data.tags.EntityTypeTagsProvider;
 import net.minecraft.world.entity.EntityType;
 import net.minecraftforge.common.data.ExistingFileHelper;
+import org.jetbrains.annotations.NotNull;
+
 import java.util.concurrent.CompletableFuture;
 
 public class DelightfulEntityTagProvider extends EntityTypeTagsProvider {
@@ -17,7 +19,7 @@ public class DelightfulEntityTagProvider extends EntityTypeTagsProvider {
 	}
 
 	@Override
-	protected void addTags(HolderLookup.Provider provider) {
+	protected void addTags(HolderLookup.@NotNull Provider provider) {
 		this.tag(DelightfulEntityTags.FATTY_ANIMALS)
 			.add(EntityType.CAT)
 			.add(EntityType.WOLF)