+++ changelog.md
@@ -1,6 +1,17 @@
 # Changelog
 
 
+# 3.5.3
+
+### Updates
+- Azalea Tea can now be made with Twigs' Azalea Flowers
+- Matcha items now use tag "forge:matcha"
+- Some items (such as teas) can be made in [Youkai's Homecoming](https://www.curseforge.com/minecraft/mc-mods/youkais-homecoming)'s kettle
+- Matcha will be disabled in the presence of [Farmer's Respite](https://www.curseforge.com/minecraft/mc-mods/farmers-respite) or [Youkai's Homecoming](https://www.curseforge.com/minecraft/mc-mods/youkais-homecoming), in favor of theirs
+- Matcha Latte's recipe is now bottled in a Honey Bottle
+- Nut Milk recipe compat for Miner's Delight's Copper Cup
+
+
 # 3.5.2
 
 ### Updates